homeWork7
