+++ random.js
@@ -1,0 +1,5 @@
+const getRandom = (num) => {
+    return Math.ceil(Math.random() * num)
+}
+
+export default getRandom
